+++ CHANGELOG.md
@@ -11,6 +11,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Name variable & code list selection is now mandatory for pairing question type
 
+## [3.3.0](https://github.com/InseeFr/Pogues/releases/tag/3.3.0) - 2026-06-11
+
+### Added
+
+- Get list of variables scopes by calling the backend
+
+### Fixed
+
+- Variables page for questionnaire version when there are scopes
+
 ## [3.2.5](https://github.com/InseeFr/Pogues/releases/tag/3.2.4) - 2026-06-09
 
 ### Added
@@ -25,9 +35,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 #### Added
 
-- Get list of variables scopes by calling the backend
 - Display scope label in variables Page
 - Default pending component (to replace by Skeleton)
+
+### Changed
+
+- Retrieve nomenclatures list from api instead of using a mock in front.
 
 ## [3.2.3](https://github.com/InseeFr/Pogues/releases/tag/3.2.3) - 2026-06-04